M-65: Moved user menu from authenticated layout to a component
