+++ resources/js/helpers/classNames.ts
@@ -1,0 +1,3 @@
+export const classNames = (...classes: string[]) => {
+	return classes.filter(Boolean).join(" ");
+};
